I might or might not have just implemented ValueTypes and unbox.any just to get 7 more successful tests
ValueTypes are unrolled (flattened) to their respective LocationsHolder. On evaluation stack they exist as plain LocationHolders. Most changes are caused by the fact that LocationDescriptor has  to remember offsets to both the primitives and refs arrays to fully mark a beginning of ValueType. A GetHashCode internalcall had to be implemented for them.